requirements.txt
beeprint==2.4.11
pyyaml==6.0.1
rich==13.7.0
argparse==1.4.0
pulp==2.9.0
flask==3.1.0